# doc/swagger/e-commerce.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "E-Commerce App",
    "version": "1.0",
    "contact": {
      "name": "kratos069",
      "url": "https://github.com/kratos069",
      "email": "dummy@email.com"
    }
  },
  "tags": [
    {
      "name": "Ecommerce"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/create_product": {
      "post": {
        "summary": "only admins are able to create a new product",
        "description": "API to create a new product",
        "operationId": "Ecommerce_CreateProduct",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateProductResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " (streaming inputs)",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateProductRequest"
            }
          }
        ],
        "tags": [
          "Ecommerce"
        ]
      }
    },
    "/v1/create_user": {
      "post": {
        "summary": "Create new user",
        "description": "API to create a new user",
        "operationId": "Ecommerce_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbCreateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbCreateUserRequest"
            }
          }
        ],
        "tags": [
          "Ecommerce"
        ]
      }
    },
    "/v1/login_user": {
      "post": {
        "summary": "login the user",
        "description": "API to login user",
        "operationId": "Ecommerce_LoginUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/pbLoginUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/pbLoginUserRequest"
            }
          }
        ],
        "tags": [
          "Ecommerce"
        ]
      }
    }
  },
  "definitions": {
    "pbCreateProductRequest": {
      "type": "object",
      "properties": {
        "info": {
          "$ref": "#/definitions/pbProductInfo"
        },
        "imageChunk": {
          "$ref": "#/definitions/pbImageChunk"
        }
      }
    },
    "pbCreateProductResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "product": {
          "$ref": "#/definitions/pbProduct"
        }
      }
    },
    "pbCreateUserRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "pbCreateUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser"
        }
      }
    },
    "pbImageChunk": {
      "type": "object",
      "properties": {
        "data": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "pbLoginUserRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "pbLoginUserResponse": {
      "type": "object",
      "properties": {
        "user": {
          "$ref": "#/definitions/pbUser"
        },
        "sessionId": {
          "type": "string"
        },
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        },
        "accessTokenExpiresAt": {
          "type": "string",
          "format": "date-time"
        },
        "refreshTokenExpiresAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "pbProduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "productImage": {
          "type": "string",
          "title": "URL from Cloudinary"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "stockQuantity": {
          "type": "string",
          "format": "int64"
        },
        "categoryId": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "pbProductInfo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "stockQuantity": {
          "type": "string",
          "format": "int64"
        },
        "categoryId": {
          "type": "string",
          "format": "int64"
        },
        "price": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "pbUser": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}